+++ docs/RELEASES.md
@@ -18,7 +18,7 @@ Legend:
 - [ ] Make wizard interactive
 
 ## <a name="7.4.0.1"></a>7.4.0.1 :ant: <sup>`1`</sup>
-- [ ] `211123>......` Fix ops page to ignore contract
+- [x] `211123>211127` Fix ops page to ignore contract
 
 ## <a name="7.4.0.0"></a>7.4.0.0 :whale: <sup>`6`</sup>
 - [x] `210911>210911` Hide closed accounts button if there are no closed accounts